requirements.txt
python-telegram-bot>=20.0
pyyaml>=6.0
python-dotenv>=1.0
aiohttp>=3.8
schedule>=1.2